Start of heat updraft system
Works, but float endlessly above lava pools, so experimental and disabled by default
To fix this I need to have an actual slowing mechanic, which means a multiplicative (not additive) value for the forward momentum, and true "air resistance", so that the player will eventually slow enough that they start falling as well. Alternatively, I could simply input the forward momentum as a value to the updraft and disable rising if they are flying below the speed of the default glider (could be subverted by holding shift all the time?). Either way, experimental for now.